requirements.txt
# local package
-e .

# external requirements
Sphinx
python-dotenv>=0.5.1
numpy
random
tqdm
collections
scipy00